PlayerMovement Implemented
The player doesn't slide anymore when there is no horizontal input. Also, player can fall faster now and a function to double jump is added which will be available to player as they progress.
Things to do next:
Player gets stuck on the side of the platform if there is horizontal input during contact.
Camera controller is needed to keep the player in the center.